feat(Radio): Replace deprecated implementation
