+++ packages/react-component-library/src/components/Fieldset/partials/StyledFieldset.tsx
@@ -3,8 +3,8 @@ import { selectors } from '@defencedigital/design-tokens'
 
 import { StyledCheckboxWrapper } from '../../Checkbox/partials/StyledCheckboxWrapper'
 import { StyledCheckbox } from '../../Checkbox/partials/StyledCheckbox'
-import { StyledRadioWrapper } from '../../RadioE/partials/StyledRadioWrapper'
-import { StyledRadio } from '../../RadioE/partials/StyledRadio'
+import { StyledRadioWrapper } from '../../Radio/partials/StyledRadioWrapper'
+import { StyledRadio } from '../../Radio/partials/StyledRadio'
 import { StyledField } from '../../Field/partials/StyledField'
 import { FieldsetType } from '../Fieldset'
