+++ packages/react-component-library/src/components/Radio/Radio.stories.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { ComponentStory, ComponentMeta } from '@storybook/react'
 
-import { Radio } from '.'
+import { RADIO_VARIANT, Radio, RadioProps } from '.'
 
 export default {
   component: Radio,
@@ -11,21 +11,41 @@ export default {
   },
 } as ComponentMeta<typeof Radio>
 
-export const Default: ComponentStory<typeof Radio> = (props) => (
-  <Radio {...props} />
-)
+const Template: ComponentStory<typeof Radio> = (props) => <Radio {...props} />
 
+const MultipleItemsTemplate: ComponentStory<typeof Radio> = (props) => {
+  function getProps(i: number): RadioProps {
+    return {
+      ...props,
+      label: `${props.label} ${i}`,
+    }
+  }
+
+  return (
+    <>
+      <Radio {...getProps(1)} />
+      <Radio {...getProps(2)} />
+      <Radio {...getProps(3)} />
+    </>
+  )
+}
+
+export const Default = Template.bind({})
 Default.args = {
   id: undefined,
   label: 'Default radio',
   name: 'default',
-  isChecked: true,
 }
 
-export const Disabled: ComponentStory<typeof Radio> = (props) => (
-  <Radio {...props} />
-)
+export const Checked = Template.bind({})
+Checked.args = {
+  id: undefined,
+  defaultChecked: true,
+  label: 'Checked radio',
+  name: 'checked',
+}
 
+export const Disabled = Template.bind({})
 Disabled.args = {
   id: undefined,
   isDisabled: true,
@@ -33,13 +53,37 @@ Disabled.args = {
   name: 'disabled',
 }
 
-export const Invalid: ComponentStory<typeof Radio> = (props) => (
-  <Radio {...props} />
-)
+export const DisabledChecked = Template.bind({})
+DisabledChecked.storyName = 'Disabled, checked'
+DisabledChecked.args = {
+  id: undefined,
+  isDisabled: true,
+  defaultChecked: true,
+  label: 'Disabled, checked radio',
+  name: 'disabled-checked',
+}
 
+export const Invalid = Template.bind({})
 Invalid.args = {
   id: undefined,
   label: 'Invalid radio',
   name: 'invalid',
   isInvalid: true,
+}
+
+export const NoContainer = MultipleItemsTemplate.bind({})
+NoContainer.args = {
+  id: undefined,
+  label: 'Item without container',
+  name: 'no-container',
+  variant: RADIO_VARIANT.NO_CONTAINER,
+}
+
+export const WithDescription = Template.bind({})
+WithDescription.args = {
+  id: undefined,
+  description:
+    'She must have hidden the plans in the escape pod. Send a detachment down to retrieve them.',
+  label: 'With description',
+  name: 'with-description',
 }
