+++ packages/react-component-library/src/components/Radio/Radio.test.tsx
@@ -1,10 +1,13 @@
 import React from 'react'
 import '@testing-library/jest-dom/extend-expect'
+import 'jest-styled-components'
+import { ColorNeutral200 } from '@defencedigital/design-tokens'
 import { render, RenderResult } from '@testing-library/react'
+import userEvent from '@testing-library/user-event'
 
+import { RADIO_VARIANT, Radio } from '.'
 import { FieldProps } from '../../common/FieldProps'
 import { FormProps } from '../../common/FormProps'
-import { Radio } from '.'
 
 describe('Radio', () => {
   let field: FieldProps
@@ -29,6 +32,8 @@ describe('Radio', () => {
   })
 
   describe('when the field has no errors, a label and a value', () => {
+    let input: HTMLElement
+
     beforeEach(() => {
       label = 'My Label 1'
       field.value = 'option1'
@@ -41,14 +46,55 @@ describe('Radio', () => {
           label={label}
         />
       )
+
+      input = radio.getByTestId('radio-input')
+    })
+
+    it('should render a container', () => {
+      expect(radio.getByTestId('radio')).toHaveStyleRule(
+        'border',
+        `1px solid ${ColorNeutral200}`
+      )
+      expect(radio.getByTestId('radio')).toHaveStyleRule(
+        'border-radius',
+        '15px'
+      )
     })
 
     it('should render a field with a label', () => {
-      expect(radio.getByTestId('label')).toHaveTextContent('My Label 1')
+      expect(radio.getByTestId('radio-label')).toHaveTextContent('My Label 1')
+    })
+
+    it('should not render a description', () => {
+      expect(radio.queryAllByTestId('checkbox-description')).toHaveLength(0)
     })
 
     it('should populate the field value', () => {
-      expect(radio.queryByTestId('radio')).toHaveAttribute('value', 'option1')
+      expect(input).toHaveAttribute('value', 'option1')
+    })
+
+    describe('and tab is pressed', () => {
+      beforeEach(() => {
+        userEvent.tab()
+      })
+
+      it('focuses the input', () => {
+        expect(input).toHaveFocus()
+      })
+
+      describe('and space is pressed', () => {
+        beforeEach(() => {
+          userEvent.keyboard('[Space]')
+        })
+
+        it('checks the input', () => {
+          expect(input).toBeChecked()
+        })
+
+        it('calls onChange once', () => {
+          expect(field.onChange).toHaveBeenCalledTimes(1)
+        })
+      })
     })
   })
 
@@ -66,7 +112,7 @@ describe('Radio', () => {
     })
 
     it('should attach the class to the wrapper', () => {
-      expect(radio.queryByTestId('container')).toHaveClass('test')
+      expect(radio.queryByTestId('radio')).toHaveClass('test')
     })
   })
 
@@ -84,11 +130,11 @@ describe('Radio', () => {
     })
 
     it('should attach the id to the field', () => {
-      expect(radio.queryByTestId('radio')).toHaveAttribute('id', 'test')
+      expect(radio.queryByTestId('radio-input')).toHaveAttribute('id', 'test')
     })
 
     it('should associate the label to the field with the custom id', () => {
-      expect(radio.queryByTestId('label')).toHaveAttribute('for', 'test')
+      expect(radio.queryByTestId('radio-label')).toHaveAttribute('for', 'test')
     })
   })
 
@@ -106,9 +152,62 @@ describe('Radio', () => {
     })
 
     it('should spread arbitrary props', () => {
-      expect(radio.getByTestId('radio')).toHaveAttribute(
+      expect(radio.getByTestId('radio-input')).toHaveAttribute(
         'data-arbitrary',
         'arbitrary'
+      )
+    })
+  })
+
+  describe('when a field has a description', () => {
+    beforeEach(() => {
+      radio = render(
+        <Radio description="Description" label="Label" name={field.name} />
+      )
+    })
+
+    it('should render a description', () => {
+      expect(radio.getByTestId('checkbox-description')).toHaveTextContent(
+        'Description'
+      )
+    })
+  })
+
+  describe('when a field has a description that is arbitrary JSX', () => {
+    beforeEach(() => {
+      radio = render(
+        <Radio
+          description={<div>Arbitrary content</div>}
+          label="Label"
+          name={field.name}
+        />
+      )
+    })
+
+    it('should render a arbitrary description', () => {
+      expect(radio.getByText('Arbitrary content')).toBeInTheDocument()
+    })
+  })
+
+  describe('when a field does not have a container', () => {
+    beforeEach(() => {
+      radio = render(
+        <Radio
+          label="Label"
+          name={field.name}
+          variant={RADIO_VARIANT.NO_CONTAINER}
+        />
+      )
+    })
+
+    it('should not render a container', () => {
+      expect(radio.getByTestId('radio')).not.toHaveStyleRule(
+        'border',
+        `1px solid ${ColorNeutral200}`
+      )
+      expect(radio.getByTestId('radio')).not.toHaveStyleRule(
+        'border-radius',
+        '15px'
       )
     })
   })
