+++ packages/react-component-library/src/components/Radio/index.ts
@@ -1,1 +1,2 @@
+export * from './constants'
 export * from './Radio'
