+++ packages/react-component-library/src/components/Radio/partials/StyledRadio.tsx
@@ -2,83 +2,138 @@ import styled, { css } from 'styled-components'
 import { selectors } from '@defencedigital/design-tokens'
 
 import { StyledCheckmark } from './StyledCheckmark'
-import { StyledLabel } from './StyledLabel'
 import { StyledInput } from './StyledInput'
 
 interface StyledRadioProps {
+  $hasContainer?: boolean
   $isDisabled?: boolean
   $isInvalid?: boolean
+  $isChecked?: boolean
 }
 
-const RADIO_FOCUS_WIDTH = '0.1rem'
+const RADIO_ACTIVE_BORDER_WIDTH = '2px'
 
 const { spacing, fontSize, color, animation } = selectors
 
+const BackgroundColor = css<StyledRadioProps>`
+  ${({ $isDisabled }) =>
+    $isDisabled ? color('neutral', '000') : color('neutral', 'white')}
+`
+
+const CheckmarkActiveBorderColor = css<StyledRadioProps>`
+  ${({ $isDisabled }) =>
+    $isDisabled ? color('neutral', '200') : color('action', '500')}
+`
+
+const CheckmarkCheckedFillColor = css<StyledRadioProps>`
+  ${({ $isDisabled }) =>
+    $isDisabled ? color('neutral', '200') : color('action', '500')}
+`
+
+function getCheckmarkCheckedSelector($hasContainer: boolean) {
+  if ($hasContainer) {
+    return css`
+      ${StyledInput}:checked ~ ${StyledCheckmark}
+    `
+  }
+
+  return css`
+    ${StyledInput}:checked ~ ${StyledCheckmark}, ${StyledInput}:focus-within ~ ${StyledCheckmark}
+  `
+}
+
 export const StyledRadio = styled.div<StyledRadioProps>`
-  display: block;
+  display: inline-flex;
   position: relative;
   padding-left: ${spacing('11')};
   cursor: pointer;
   font-size: ${fontSize('base')};
   user-select: none;
+  background: ${BackgroundColor};
 
-  /* Active state, light blue outline */
-  ${StyledInput}:hover ~ ${StyledCheckmark},
-  ${StyledInput}:active ~ ${StyledCheckmark} {
-    border: 1px solid ${color('action', '500')};
-    outline: none;
-    box-shadow: 0 0 0 ${RADIO_FOCUS_WIDTH} ${color('action', '500')};
-    transition: all ${animation('default')};
+  * {
+    cursor: pointer;
   }
 
-  /* Checked state, blue border */
-  ${StyledInput}:checked ~ ${StyledCheckmark} {
+  ${StyledCheckmark} {
+    background: ${BackgroundColor};
+
+    &::before {
+      display: block;
+      box-shadow: 0 0 0 0 transparent;
+      transition: all ${animation('default')};
+    }
+  }
+
+  ${({ $hasContainer }) =>
+    $hasContainer &&
+    css`
+      border: 1px solid ${color('neutral', '200')};
+      border-radius: 15px;
+
+      &:focus-within,
+      &:active {
+        outline: none;
+        border-color: ${color('action', '500')};
+        box-shadow: 0 0 0 2px ${color('action', '500')},
+          0 0 0 5px ${color('action', '100')};
+      }
+    `}
+
+  /* Checkmark hover and active states, blue border */
+
+  ${({ $isDisabled }) =>
+    !$isDisabled &&
+    css`
+      &:hover ${StyledCheckmark}, ${StyledInput}:active ~ ${StyledCheckmark} {
+        &::before {
+          box-shadow: 0 0 0 ${RADIO_ACTIVE_BORDER_WIDTH}
+            ${color('action', '500')};
+        }
+      }
+    `}
+
+  /* Checkmark checked state */
+
+  ${({ $hasContainer }) => getCheckmarkCheckedSelector($hasContainer)} {
     background-color: ${color('neutral', 'white')};
-    border: 1px solid ${color('action', '500')};
-    box-shadow: 0 0 0 ${RADIO_FOCUS_WIDTH} ${color('action', '500')};
-  }
 
-  ${StyledInput}:checked ~ ${StyledCheckmark}:after {
-    display: block;
-  }
+    /* Blue border (grey if disabled) */
+    &::before {
+      box-shadow: 0 0 0 ${RADIO_ACTIVE_BORDER_WIDTH}
+        ${CheckmarkActiveBorderColor};
+    }
 
-  /* Checkmark, blue dot */
-  ${StyledCheckmark}::after {
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    border-radius: 999px;
-    background: ${color('action', '600')};
-    border: 3px solid ${color('neutral', 'white')};
+    /* Central blue dot (grey dot if disabled) */
+    &::after {
+      display: block;
+      background: ${CheckmarkCheckedFillColor};
+      border: 2px solid ${BackgroundColor};
+    }
   }
 
   ${({ $isDisabled }) =>
     $isDisabled &&
     css`
-      ${StyledLabel} {
+      cursor: not-allowed;
+
+      * {
         cursor: not-allowed;
-        color: ${color('neutral', '200')};
       }
 
-      ${StyledInput} ~ ${StyledCheckmark} {
-        background-color: ${color('neutral', '000')};
-        border: 1px solid ${color('neutral', '100')};
-      }
+      border-color: transparent;
 
-      ${StyledInput}:hover ~ ${StyledCheckmark},
-      ${StyledInput}:active ~ .rn-radio__checkmark {
-        border: 1px solid ${color('neutral', '100')};
+      &:focus-within,
+      &:active {
+        border-color: transparent;
         box-shadow: none;
-        cursor: not-allowed;
       }
     `}
 
   ${({ $isInvalid }) =>
     $isInvalid &&
     css`
-      ${StyledLabel} {
-        color: ${color('danger', '700')};
-      }
+      border-color: ${color('danger', '800')};
+      box-shadow: 0 0 0 2px ${color('danger', '800')};
     `}
 `
