--- packages/react-component-library/src/components/RadioE/index.ts
@@ -1,2 +1,0 @@
-export * from './constants'
-export * from './RadioE'
