--- packages/react-component-library/src/components/RadioE/partials/StyledInput.tsx
@@ -1,9 +1,0 @@
-import styled from 'styled-components'
-
-export const StyledInput = styled.input`
-  position: absolute;
-  opacity: 0;
-  cursor: pointer;
-  height: 0;
-  width: 0;
-`
