--- packages/react-component-library/src/components/RadioEnhanced/index.ts
@@ -1,1 +1,0 @@
-export * from './RadioEnhanced'
