--- packages/react-component-library/src/components/RadioEnhanced/partials/StyledDescription.tsx
@@ -1,9 +1,0 @@
-import styled from 'styled-components'
-import { selectors } from '@defencedigital/design-tokens'
-
-const { color, spacing } = selectors
-
-export const StyledDescription = styled.p`
-  color: ${color('neutral', '400')};
-  padding-top: ${spacing('4')};
-`
