+++ packages/react-component-library/src/forms/native/native.stories.tsx
@@ -4,7 +4,7 @@ import { ComponentMeta } from '@storybook/react'
 
 import { TextInputE } from '../../components/TextInputE'
 import { TextAreaE } from '../../components/TextAreaE'
-import { RadioE } from '../../components/RadioE'
+import { Radio } from '../../components/Radio'
 import {
   AutocompleteE,
   AutocompleteEOption,
@@ -139,13 +139,13 @@ export const Example: React.FC<unknown> = () => {
           />
         </Fieldset>
         <Fieldset legend="Example radio selection">
-          <RadioE
+          <Radio
             onChange={handleRadioGroup}
             name="exampleRadio"
             label="Option 1"
             value="Option 1"
           />
-          <RadioE
+          <Radio
             onChange={handleRadioGroup}
             name="exampleRadio"
             label="Option 2"
