+++ packages/react-component-library/src/index.ts
@@ -39,8 +39,6 @@ export * from './components/PhaseBanner'
 export * from './components/Popover'
 export * from './components/ProgressIndicator'
 export * from './components/Radio'
-export * from './components/RadioE'
-export * from './components/RadioEnhanced'
 export * from './components/RangeSlider'
 export * from './components/RangeSliderE'
 export * from './components/SectionDivider'
